fix: pessimistically lookup the earliest key
The earliest_offset function in the consumer seems to not always return
all tps which have been passed in but presumabely only the ones it has
data for.

After longer runtime it seems that the response MAY return a dict
not filled with all topic partitions which have been added. To bypass
this and not lead faust to crash, use the .get function on the returned
mapping and default to None on return.